refactor: Create Category table
Create a table for the product Category. The Category now has a CategoryId and Name (which must be unique). The Product table now uses a CategoryId.

Getting this to work required removing the database tables and reverting the migrations. Fortunately, having the Seed function meant not having to manually re-introduce data.