refactor: Implement partial views for displaying product lists and details
Abstract the code for displaying a list of products with full info (_ProductList) and displaying a product's details (_ProductDetail) into partial views. The former is used in Index, which is also used by the Search function; the latter is used by both Details and Delete.

Also define the _products table in ProductsController. This table is an eager join with the Category table so that the item category is available for all views.